Tweaked web3 button to allow for use with contract interaction
