+++ src/components/Web3Button.jsx
@@ -1,8 +1,11 @@
 import React from "react";
 
+import { useWeb3Context } from 'web3-react'
+
 export default function Web3Button(props) {
-  if (props.show) {
-    return <button onClick={props.onClick}>{props.text}</button>;
+  const context = useWeb3Context()
+  if (props.show(context)) {
+    return <button onClick={() => props.onClick(context)}>{props.text}</button>;
   } else {
     return null;
   }
